Move all command scripts in same folder
